refactor: movendo as exceptions pra infrastructure
