+++ src/application/usecases/auth/user-login.use-case.ts
@@ -2,7 +2,7 @@ import type IUseCases from '@/application/usecases/interfaces/use-cases.interfac
 import type UserLoginEntity from '@/domain/user/entities/user-login.entity'
 import type UserEntity from '@/domain/user/entities/user.entity'
 import type IUserRepository from '@/domain/user/interfaces/user-repository.interface'
-import { UsuarioNaoEncontradoException } from '@/shared/common/exceptions/user/usuario-nao-encontrado.exception'
+import { UsuarioNaoEncontradoException } from '@/infrastructure/exceptions/user/usuario-nao-encontrado.exception'
 
 export default class UserLoginUseCase implements IUseCases {
   constructor(private iUserRepository: IUserRepository<UserEntity>) {}
