+++ src/application/usecases/bakery/product/remove-product.use-case.ts
@@ -3,8 +3,8 @@ import type BakeryManagerEntity from '@/domain/bakery/entities/bakery-manager.en
 import type ProductEntity from '@/domain/bakery/entities/product.entity'
 import type IBakeryManagerRepository from '@/domain/bakery/interfaces/bakery-manager-repository.interface'
 import type IProductRepository from '@/domain/bakery/interfaces/product.repository'
-import { ProdutoNaoEncontradoException } from '@/shared/common/exceptions/bakery/product/produto-nao-encontrado.exception'
-import { UsuarioNaoEGerenteException } from '@/shared/common/exceptions/bakery/product/usuario-nao-gerente.exception'
+import { ProdutoNaoEncontradoException } from '@/infrastructure/exceptions/bakery/product/produto-nao-encontrado.exception'
+import { UsuarioNaoEGerenteException } from '@/infrastructure/exceptions/bakery/product/usuario-nao-gerente.exception'
 
 export default class RemoveProductUseCase implements IUseCases {
   constructor(
