+++ src/application/usecases/user/create-user.use-case.ts
@@ -1,7 +1,7 @@
 import type IUseCases from '@/application/usecases/interfaces/use-cases.interface'
 import type UserEntity from '@/domain/user/entities/user.entity'
 import type IUserRepository from '@/domain/user/interfaces/user-repository.interface'
-import { EmailOuUsernameExistenteException } from '@/shared/common/exceptions/user/email-ou-username-existente.exception'
+import { EmailOuUsernameExistenteException } from '@/infrastructure/exceptions/user/email-ou-username-existente.exception'
 
 export default class CreateUserUseCase implements IUseCases {
   constructor(private iUserRepository: IUserRepository<UserEntity>) {}
