+++ src/application/usecases/user/find-one-user.use-case.ts
@@ -1,7 +1,7 @@
 import type IUseCases from '@/application/usecases/interfaces/use-cases.interface'
 import type UserEntity from '@/domain/user/entities/user.entity'
 import type IUserRepository from '@/domain/user/interfaces/user-repository.interface'
-import { UsuarioNaoEncontradoException } from '@/shared/common/exceptions/user/usuario-nao-encontrado.exception'
+import { UsuarioNaoEncontradoException } from '@/infrastructure/exceptions/user/usuario-nao-encontrado.exception'
 
 export default class FindOneUserUseCase implements IUseCases {
   constructor(private iUserRepository: IUserRepository<UserEntity>) {}
