+++ src/infrastructure/exceptions/bakery/bakery-manager/padaria-ja-possui-este-gerente.exception.ts
