+++ src/infrastructure/exceptions/bakery/bakery-manager/relacao-entre-padaria-e-gerente-nao-encontrada.exception.ts
