+++ src/infrastructure/exceptions/bakery/padaria-nao-encontrada.exception.ts
