+++ src/infrastructure/exceptions/bakery/product/produto-nao-encontrado.exception.ts
