+++ src/infrastructure/exceptions/bakery/product/usuario-nao-gerente.exception.ts
