+++ src/infrastructure/exceptions/user/autenticacao-nao-efetuada.exception.ts
