+++ src/infrastructure/exceptions/user/gerente-nao-encontrado.exception.ts
