+++ src/infrastructure/exceptions/user/nenhum-gerente-encontrado.exception.ts
