+++ src/infrastructure/exceptions/user/senha-invalida.exception.ts
