+++ src/infrastructure/exceptions/user/user-role-vazia.exception.ts
