+++ src/infrastructure/exceptions/user/usuario-nao-encontrado.exception.ts
